backend/ai-engine/requirements.txt
fastapi
uvicorn
pydantic
motor
python-jose[cryptography]
passlib[bcrypt]
google-genai
python-dotenv
email-validator
